+++ src/Dashboard/Dashboard.tsx
@@ -1,27 +1,16 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { getRandomCocktail } from '../apiCalls';
 import './Dashboard.scss';
 import { Cocktail } from '../Definitions/RandomCocktail'
 
 export interface DashboardProps {
-	username: string
+	username: string;
+	randomCocktail: Cocktail
 }
 
 const Dashboard: React.SFC<DashboardProps> = (props) => {
-	const [randomCocktail, setRandomCocktail] = useState<Cocktail>({idDrink: '', strDrink: '', strInstructions: '', strDrinkThumb: ''});
+	const [randomCocktail, setRandomCocktail] = useState<Cocktail>(props.randomCocktail);
 	const [error, setError] = useState<string>('');
-
-	const getCocktail = async ():Promise<any> => {
-		try {
-			const data: Cocktail = await getRandomCocktail();
-			setRandomCocktail(data);
-		} catch (error) {
-			setError(error.message);
-		}
-	}
-
-	useEffect(() => {getCocktail()}, [])
 
 	return (
 		<section className='cocktail-OTD-wrapper'>
