+++ src/MyCocktails/MyCocktails.tsx
@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import { getAllCocktails } from '../apiCalls';
 import './MyCocktails.scss';
 
 interface MyCocktailsProps {
